test(runtime): isolate event-loop control-flow regressions
