+++ tests/CMakeLists.txt
@@ -5530,6 +5530,7 @@ copperfin_add_test(test_prg_engine_control_flow cf_xbase_runtime
     test_prg_engine_control_flow_aggregate_and_total_commands.cpp
     test_prg_engine_select_query_commands.cpp
     test_prg_engine_control_flow_concurrency_and_critical_sections.cpp
+    test_prg_engine_control_flow_event_loop.cpp
     test_prg_engine_control_flow_error_handler_commands.cpp
     test_prg_engine_control_flow_bare_null.cpp
     test_prg_engine_control_flow_on_error_dispatch.cpp
